backend/requirements.txt
flask
pydantic
PyGithub
python-dotenv
flask-cors
requests
redis
rq